requirements.txt
paho-mqtt==1.4.0
Adafruit-DHT==1.4.0
RPi.bme280==0.2.3
